requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2024.2
dj-database-url==2.2.0
Django==5.0.6
django-crispy-forms==2.1
mysqlclient==2.2.4
pillow==10.4.0
python-dotenv==1.0.1
sqlparse==0.5.0
typing_extensions==4.12.1
tzdata==2024.1
